특수기호 사전 정비 및 추가
