+++ data/hanja/emoticon.txt
@@ -1,0 +1,846 @@
+ㅏ:😄:웃는 표정
+ㅏ:😃:웃는 표정
+ㅏ:😀:웃는 표정
+ㅏ:😊:웃는 표정
+ㅏ:☺️:웃는 표정
+ㅏ:😉:윙크하는 표정
+ㅏ:😍:하트모양 눈 표정
+ㅏ:😘:키스 날리는 표정
+ㅏ:😚:키스하는 표정
+ㅏ:😗:키스하는 표정
+ㅏ:😙:키스하는 표정
+ㅏ:😜:혀 빼고 윙크하는 표정
+ㅏ:😝:혀 빼고 눈 꼭 감은 표정
+ㅏ:😛:혀 빼고 있는 표정
+ㅏ:😳:상기된 표정
+ㅏ:😁:이 보이며 웃는 표정
+ㅏ:😔:생각에 잠긴 표정
+ㅏ:😌:편안한 표정
+ㅏ:😒:지루한 표정
+ㅏ:😞:실망한 표정
+ㅏ:😣:참는 표정
+ㅏ:😢:우는 표정
+ㅏ:😂:웃겨서 눈물나는 표정
+ㅏ:😭:시끄럽게 우는 표정
+ㅏ:😪:졸리는 표정
+ㅏ:😥:실망했지만 편안한 표정
+ㅏ:😰:입 벌리고 식은땀 흘리는 표정
+ㅏ:😅:식은땀 흘리며 웃는 표정
+ㅏ:😓:식은땀 흘리는 표정
+ㅏ:😩:피곤한 표정
+ㅏ:😫:표정
+ㅏ:😨:표정
+ㅏ:😱:표정
+ㅏ:😠:표정
+ㅏ:😡:표정
+ㅏ:😤:표정
+ㅏ:😖:표정
+ㅏ:😆:표정
+ㅏ:😋:표정
+ㅏ:😷:표정
+ㅏ:😎:표정
+ㅏ:😴:표정
+ㅏ:😵:표정
+ㅏ:😲:표정
+ㅏ:😟:표정
+ㅏ:😦:표정
+ㅏ:😧:표정 
+ㅏ:😈:표정 
+ㅏ:👿:표정 
+ㅏ:😮:표정 
+ㅏ:😬:표정 
+ㅏ:😐:표정 
+ㅏ:😕:표정 
+ㅏ:😯:표정 
+ㅏ:😶:표정 
+ㅏ:😇:표정 
+ㅏ:😏:표정 
+ㅏ:😑:표정 
+ㅏ:👲:사람
+ㅏ:👳:사람
+ㅏ:👮:사람
+ㅏ:👷:사람
+ㅏ:💂:사람
+ㅏ:👶:사람
+ㅏ:👦:사람
+ㅏ:👧:사람
+ㅏ:👨:사람
+ㅏ:👩:사람
+ㅏ:👴:사람
+ㅏ:👵:사람
+ㅏ:👱:사람
+ㅏ:👼:사람
+ㅏ:👸:사람
+ㅏ:😺:고양이
+ㅏ:😸:고양이
+ㅏ:😻:고양이
+ㅏ:😽:고양이
+ㅏ:😼:고양이
+ㅏ:🙀:고양이
+ㅏ:😿:고양이
+ㅏ:😹:고양이
+ㅏ:😾:고양이
+ㅏ:👹:일본 도깨비
+ㅏ:👺:일본 도깨비
+ㅏ:🙈:원숭이
+ㅏ:🙉:원숭이
+ㅏ:🙊:원숭이
+ㅏ:💀:해골
+ㅏ:👽:외계인
+ㅏ:💩:똥무더기
+ㅏ:🔥:불
+ㅏ:✨:빛
+ㅏ:🌟:별
+ㅏ:💫:어지러움
+ㅏ:💥:충돌
+ㅏ:💢:화남
+ㅏ:💦:흐르는 땀
+ㅏ:💧:물방울
+ㅏ:💤:잠자는 중
+ㅏ:💨:달리기
+ㅏ:👂:귀
+ㅏ:👀:눈
+ㅏ:👃:코
+ㅏ:👅:혀
+ㅏ:👄:입술
+ㅏ:👍:손가락 올림
+ㅏ:👎:손가락 내림
+ㅏ:👌:확인 수신호
+ㅏ:👊:내지른 주먹
+ㅏ:✊:들어올린 주먹
+ㅏ:✌️:승리 수신호
+ㅏ:👋:손흔드는 신호
+ㅏ:✋:들어올린 손
+ㅏ:👐:활짝 벌린 손
+ㅏ:👆:위를 가리키는 손등
+ㅏ:👇:아래를 가리키는손등
+ㅏ:👉:오른쪽을 가리키는 손등
+ㅏ:👈:왼쪽을 가리키는 손등
+ㅏ:🙌:양손 만세
+ㅏ:🙏:손뼉치는사람
+ㅏ:☝️:위를 가리키는 손
+ㅏ:👏:손뼉 신호
+ㅏ:💪:이두박근
+ㅏ:🚶:사람
+ㅏ:🏃:사람
+ㅏ:💃:사람
+ㅏ:👫:사람
+ㅏ:👪:사람
+ㅏ:👬:사람
+ㅏ:👭:사람
+ㅏ:💏:사람
+ㅏ:💑:사람
+ㅏ:👯:사람
+ㅏ:🙆:사람
+ㅏ:🙅:사람
+ㅏ:💁:사람
+ㅏ:🙋:사람
+ㅏ:💆:사람
+ㅏ:💇:사람
+ㅏ:💅:사람
+ㅏ:👰:사람
+ㅏ:🙎:사람
+ㅏ:🙍:사람
+ㅏ:🙇:사람
+ㅏ:🎩:영국 모자
+ㅏ:👑:왕관
+ㅏ:👒:부인 모자
+ㅏ:👟:운동화
+ㅏ:👞:남자 구두
+ㅏ:👡:여자 샌달
+ㅏ:👠:하이힐
+ㅏ:👢:여자 부츠
+ㅏ:👕:티셔츠
+ㅏ:👔:넥타이
+ㅏ:👚:여자옷
+ㅏ:👗:치마
+ㅏ:🎽:어깨끈 달린 운동복
+ㅏ:👖:청바지
+ㅏ:👘:기모노
+ㅏ:👙:비키니
+ㅏ:💼:서류가방
+ㅏ:👜:손가방
+ㅏ:👝:작은 손가방
+ㅏ:👛:지갑
+ㅏ:👓:안경
+ㅏ:🎀:리본
+ㅏ:🌂:우산
+ㅏ:💄:립스틱
+ㅏ:💛:하트
+ㅏ:💙:하트
+ㅏ:💜:하트
+ㅏ:💚:하트
+ㅏ:❤️:하트
+ㅏ:💔:깨진 하트
+ㅏ:💗:하트
+ㅏ:💓:하트
+ㅏ:💕:하트
+ㅏ:💖:하트
+ㅏ:💞:하트
+ㅏ:💘:하트
+ㅏ:💌:러브레터
+ㅏ:💋:키스마크
+ㅏ:💍:반지
+ㅏ:💎:보석
+ㅏ:👤:사람
+ㅏ:👥:사람
+ㅏ:💬:말풍선
+ㅏ:👣:발자국
+ㅏ:💭:생각풍선
+ㅓ:🐶:동물
+ㅓ:🐺:동물
+ㅓ:🐱:동물
+ㅓ:🐭:동물
+ㅓ:🐹:동물
+ㅓ:🐰:동물
+ㅓ:🐸:동물
+ㅓ:🐯:동물
+ㅓ:🐨:동물
+ㅓ:🐻:동물
+ㅓ:🐷:동물
+ㅓ:🐽:동물
+ㅓ:🐮:동물
+ㅓ:🐗:동물
+ㅓ:🐵:동물
+ㅓ:🐒:동물
+ㅓ:🐴:동물
+ㅓ:🐑:동물
+ㅓ:🐘:동물
+ㅓ:🐼:동물
+ㅓ:🐧:동물
+ㅓ:🐦:동물
+ㅓ:🐤:동물
+ㅓ:🐥:동물
+ㅓ:🐣:동물
+ㅓ:🐔:동물
+ㅓ:🐍:동물
+ㅓ:🐢:동물
+ㅓ:🐛:동물
+ㅓ:🐝:동물
+ㅓ:🐜:동물
+ㅓ:🐞:동물
+ㅓ:🐌:동물
+ㅓ:🐙:동물
+ㅓ:🐚:동물
+ㅓ:🐠:동물
+ㅓ:🐟:동물
+ㅓ:🐬:동물
+ㅓ:🐳:동물
+ㅓ:🐋:동물
+ㅓ:🐄:동물
+ㅓ:🐏:동물
+ㅓ:🐀:동물
+ㅓ:🐃:동물
+ㅓ:🐅:동물
+ㅓ:🐇:동물
+ㅓ:🐉:동물
+ㅓ:🐎:동물
+ㅓ:🐐:동물
+ㅓ:🐓:동물
+ㅓ:🐕:동물
+ㅓ:🐖:동물
+ㅓ:🐁:동물
+ㅓ:🐂:동물
+ㅓ:🐲:동물
+ㅓ:🐡:동물
+ㅓ:🐊:동물
+ㅓ:🐫:동물
+ㅓ:🐪:동물
+ㅓ:🐆:동물
+ㅓ:🐈:동물
+ㅓ:🐩:동물
+ㅓ:🐾:발자국
+ㅓ:💐:식물
+ㅓ:🌸:식물
+ㅓ:🌷:식물
+ㅓ:🍀:식물
+ㅓ:🌹:식물
+ㅓ:🌻:식물
+ㅓ:🌺:식물
+ㅓ:🍁:식물
+ㅓ:🍃:식물
+ㅓ:🍂:식물
+ㅓ:🌿:식물
+ㅓ:🌾:식물
+ㅓ:🍄:식물
+ㅓ:🌵:식물
+ㅓ:🌴:식물
+ㅓ:🌲:식물
+ㅓ:🌳:식물
+ㅓ:🌰:밤
+ㅓ:🌱:새싹
+ㅓ:🌼:활짝 핀 꽃
+ㅓ:🌐:천체
+ㅓ:🌞:천체
+ㅓ:🌝:천체
+ㅓ:🌚:천체
+ㅓ:🌑:천체
+ㅓ:🌒:천체
+ㅓ:🌓:천체
+ㅓ:🌔:천체
+ㅓ:🌕:천체
+ㅓ:🌖:천체
+ㅓ:🌗:천체
+ㅓ:🌘:천체
+ㅓ:🌜:천체
+ㅓ:🌛:천체
+ㅓ:🌙:천체
+ㅓ:🌍:천체
+ㅓ:🌎:천체
+ㅓ:🌏:천체
+ㅓ:🌋:천체
+ㅓ:🌌:천체
+ㅓ:🌠:천체
+ㅓ:⭐:천체
+ㅓ:☀️:날씨
+ㅓ:⛅:날씨
+ㅓ:☁️:날씨
+ㅓ:⚡️:날씨
+ㅓ:☔️:날씨
+ㅓ:❄️:날씨
+ㅓ:⛄:날씨
+ㅓ:🌀:날씨
+ㅓ:🌁:날씨
+ㅓ:🌈:날씨
+ㅓ:🌊:날씨
+ㅗ:🎍:사물
+ㅗ:💝:사물
+ㅗ:🎎:사물
+ㅗ:🎒:사물
+ㅗ:🎓:사물
+ㅗ:🎏:사물
+ㅗ:🎆:사물
+ㅗ:🎇:사물
+ㅗ:🎐:사물
+ㅗ:🎑:사물
+ㅗ:🎃:사물
+ㅗ:👻:사물
+ㅗ:🎅:사물
+ㅗ:🎄:사물
+ㅗ:🎁:사물
+ㅗ:🎋:사물
+ㅗ:🎉:사물
+ㅗ:🎊:사물
+ㅗ:🎈:사물
+ㅗ:🎌:사물
+ㅗ:🔮:사물
+ㅗ:🎥:사물
+ㅗ:📷:사물
+ㅗ:📹:사물
+ㅗ:📼:사물
+ㅗ:💿:사물
+ㅗ:📀:사물
+ㅗ:💽:사물
+ㅗ:💾:사물
+ㅗ:💻:사물
+ㅗ:📱:사물
+ㅗ:☎:사물
+ㅗ:📞:사물
+ㅗ:📟:사물
+ㅗ:📠:사물
+ㅗ:📡:사물
+ㅗ:📺:사물
+ㅗ:📻:사물
+ㅗ:🔊:사물
+ㅗ:🔉:사물
+ㅗ:🔈:사물
+ㅗ:🔇:사물
+ㅗ:🔔:사물
+ㅗ:🔕:사물
+ㅗ:📢:사물
+ㅗ:📣:사물
+ㅗ:⏳:사물
+ㅗ:⌛:사물
+ㅗ:⏰:사물
+ㅗ:⌚:사물
+ㅗ:🔓:사물
+ㅗ:🔒:사물
+ㅗ:🔏:사물
+ㅗ:🔐:사물
+ㅗ:🔑:사물
+ㅗ:🔎:사물
+ㅗ:💡:사물
+ㅗ:🔦:사물
+ㅗ:🔆:사물
+ㅗ:🔅:사물
+ㅗ:🔌:사물
+ㅗ:🔋:사물
+ㅗ:🔍:사물
+ㅗ:🛁:사물
+ㅗ:🛀:사물
+ㅗ:🚿:사물
+ㅗ:🚽:사물
+ㅗ:🔧:사물
+ㅗ:🔩:사물
+ㅗ:🔨:사물
+ㅗ:🚪:사물
+ㅗ:🚬:사물
+ㅗ:💣:사물
+ㅗ:🔫:사물
+ㅗ:🔪:사물
+ㅗ:💊:사물
+ㅗ:💉:사물
+ㅗ:💰:사물
+ㅗ:💴:사물
+ㅗ:💵:사물
+ㅗ:💷:사물
+ㅗ:💶:사물
+ㅗ:💳:사물
+ㅗ:💸:사물
+ㅗ:📲:사물
+ㅗ:📧:사물
+ㅗ:📥:사물
+ㅗ:📤:사물
+ㅗ:✉:사물
+ㅗ:📩:사물
+ㅗ:📨:사물
+ㅗ:📯:사물
+ㅗ:📫:사물
+ㅗ:📪:사물
+ㅗ:📬:사물
+ㅗ:📭:사물
+ㅗ:📮:사물
+ㅗ:📦:사물
+ㅗ:📝:사물
+ㅗ:📄:사물
+ㅗ:📃:사물
+ㅗ:📑:사물
+ㅗ:📊:사물
+ㅗ:📈:사물
+ㅗ:📉:사물
+ㅗ:📜:사물
+ㅗ:📋:사물
+ㅗ:📅:사물
+ㅗ:📆:사물
+ㅗ:📇:사물
+ㅗ:📁:사물
+ㅗ:📂:사물
+ㅗ:✂:사물
+ㅗ:📌:사물
+ㅗ:📎:사물
+ㅗ:✒:사물
+ㅗ:✏:사물
+ㅗ:📏:사물
+ㅗ:📐:사물
+ㅗ:📕:사물
+ㅗ:📗:사물
+ㅗ:📘:사물
+ㅗ:📙:사물
+ㅗ:📓:사물
+ㅗ:📔:사물
+ㅗ:📒:사물
+ㅗ:📚:사물
+ㅗ:📖:사물
+ㅗ:🔖:사물
+ㅗ:📛:사물
+ㅗ:🔬:사물
+ㅗ:🔭:사물
+ㅗ:📰:사물
+ㅗ:🎨:사물
+ㅗ:🎬:사물
+ㅗ:🎤:사물
+ㅗ:🎧:사물
+ㅗ:🎼:사물
+ㅗ:🎵:사물
+ㅗ:🎶:사물
+ㅗ:🎹:사물
+ㅗ:🎻:사물
+ㅗ:🎺:사물
+ㅗ:🎷:사물
+ㅗ:🎸:사물
+ㅗ:👾:사물
+ㅗ:🎮:사물
+ㅗ:🃏:사물
+ㅗ:🎴:사물
+ㅗ:🀄:사물
+ㅗ:🎲:사물
+ㅗ:🎯:사물
+ㅗ:🏈:사물
+ㅗ:🏀:사물
+ㅗ:⚽:사물
+ㅗ:⚾:사물
+ㅗ:🎾:사물
+ㅗ:🎱:사물
+ㅗ:🏉:사물
+ㅗ:🎳:사물
+ㅗ:⛳:사물
+ㅗ:🚵:사물
+ㅗ:🚴:사물
+ㅗ:🏁:사물
+ㅗ:🏇:사물
+ㅗ:🏆:사물
+ㅗ:🎿:사물
+ㅗ:🏂:사물
+ㅗ:🏊:사물
+ㅗ:🏄:사물
+ㅗ:🎣:사물
+ㅗ:☕:사물
+ㅗ:🍵:사물
+ㅗ:🍶:사물
+ㅗ:🍼:사물
+ㅗ:🍺:사물
+ㅗ:🍻:사물
+ㅗ:🍸:사물
+ㅗ:🍹:사물
+ㅗ:🍷:사물
+ㅗ:🍴:사물
+ㅗ:🍕:사물
+ㅗ:🍔:사물
+ㅗ:🍟:사물
+ㅗ:🍗:사물
+ㅗ:🍖:사물
+ㅗ:🍝:사물
+ㅗ:🍛:사물
+ㅗ:🍤:사물
+ㅗ:🍱:사물
+ㅗ:🍣:사물
+ㅗ:🍥:사물
+ㅗ:🍙:사물
+ㅗ:🍘:사물
+ㅗ:🍚:사물
+ㅗ:🍜:사물
+ㅗ:🍲:사물
+ㅗ:🍢:사물
+ㅗ:🍡:사물
+ㅗ:🍳:사물
+ㅗ:🍞:사물
+ㅗ:🍩:사물
+ㅗ:🍮:사물
+ㅗ:🍦:사물
+ㅗ:🍨:사물
+ㅗ:🍧:사물
+ㅗ:🎂:사물
+ㅗ:🍰:사물
+ㅗ:🍪:사물
+ㅗ:🍫:사물
+ㅗ:🍬:사물
+ㅗ:🍭:사물
+ㅗ:🍯:사물
+ㅗ:🍎:사물
+ㅗ:🍏:사물
+ㅗ:🍊:사물
+ㅗ:🍋:사물
+ㅗ:🍒:사물
+ㅗ:🍇:사물
+ㅗ:🍉:사물
+ㅗ:🍓:사물
+ㅗ:🍑:사물
+ㅗ:🍈:사물
+ㅗ:🍌:사물
+ㅗ:🍐:사물
+ㅗ:🍍:사물
+ㅗ:🍠:사물
+ㅗ:🍆:사물
+ㅗ:🍅:사물
+ㅗ:🌽:사물
+ㅜ:🏠:건물
+ㅜ:🏡:건물
+ㅜ:🏫:건물
+ㅜ:🏢:건물
+ㅜ:🏣:건물
+ㅜ:🏥:건물
+ㅜ:🏦:건물
+ㅜ:🏪:건물
+ㅜ:🏩:건물
+ㅜ:🏨:건물
+ㅜ:💒:건물
+ㅜ:⛪:건물
+ㅜ:🏬:건물
+ㅜ:🏤:건물
+ㅜ:🌇:건물
+ㅜ:🌆:건물
+ㅜ:🏯:건물
+ㅜ:🏰:건물
+ㅜ:⛺:건물
+ㅜ:🏭:건물
+ㅜ:🗼:건물
+ㅜ:🗾:일본 지도
+ㅜ:🗻:후지산
+ㅜ:🌄:해돋이
+ㅜ:🌅:해몰이
+ㅜ:🌃:별 뜬 야경
+ㅜ:🗽:자유의 여신상
+ㅜ:🌉:다리 야경
+ㅜ:🎠:회전목마
+ㅜ:🎡:관람차
+ㅜ:⛲:분수
+ㅜ:🎢:롤러코스터
+ㅜ:🚢:항해
+ㅜ:⛵:항해
+ㅜ:🚤:항해
+ㅜ:🚣:항해
+ㅜ:⚓️:항해
+ㅜ:🚀:비행
+ㅜ:✈️:비행
+ㅜ:💺:비행
+ㅜ:🚁:비행
+ㅜ:🚂:철도
+ㅜ:🚊:철도
+ㅜ:🚉:철도
+ㅜ:🚞:철도
+ㅜ:🚆:철도
+ㅜ:🚄:철도
+ㅜ:🚅:철도
+ㅜ:🚈:철도
+ㅜ:🚇:철도
+ㅜ:🚝:철도
+ㅜ:🚋:철도
+ㅜ:🚃:철도
+ㅜ:🚎:철도
+ㅜ:🚌:자동차
+ㅜ:🚍:자동차
+ㅜ:🚙:자동차
+ㅜ:🚘:자동차
+ㅜ:🚗:자동차
+ㅜ:🚕:자동차
+ㅜ:🚖:자동차
+ㅜ:🚛:자동차
+ㅜ:🚚:자동차
+ㅜ:🚨:자동차
+ㅜ:🚓:자동차
+ㅜ:🚔:자동차
+ㅜ:🚒:자동차
+ㅜ:🚑:자동차
+ㅜ:🚐:자동차
+ㅜ:🚲:자전거
+ㅜ:🚡:삭도차
+ㅜ:🚟:삭도차
+ㅜ:🚠:삭도차
+ㅜ:🚜:트랙터
+ㅜ:💈:이발소
+ㅜ:🚏:버스 정류장
+ㅜ:🎫:차표
+ㅜ:🚦:신호등
+ㅜ:🚥:신호등
+ㅜ:⚠️:경고 표지판
+ㅜ:🚧:공사 표지판
+ㅜ:🔰:기호
+ㅜ:⛽:기호
+ㅜ:🏮:기호
+ㅜ:🎰:기호
+ㅜ:♨️:기호
+ㅜ:🗿:모아이 석상
+ㅜ:🎪:서커스 천막
+ㅜ:🎭:공연 예술
+ㅜ:📍:둥근 핀
+ㅜ:🚩:삼각 깃발
+ㅜ:🇯🇵:일본 국기
+ㅜ:🇰🇷:한국 국기
+ㅜ:🇩🇪:독일 국기
+ㅜ:🇨🇳:중국 국기
+ㅜ:🇺🇸:미국 국기
+ㅜ:🇫🇷:프랑스 국기
+ㅜ:🇪🇸:스페인 국기
+ㅜ:🇮🇹:이탈리아 국기
+ㅜ:🇷🇺:러시아 국기
+ㅜ:🇬🇧:영국 국기
+# 기호
+ㅡ:1️⃣:숫자 기호
+ㅡ:2️⃣:숫자 기호
+ㅡ:3️⃣:숫자 기호
+ㅡ:4️⃣:숫자 기호
+ㅡ:5️⃣:숫자 기호
+ㅡ:6️⃣:숫자 기호
+ㅡ:7️⃣:숫자 기호
+ㅡ:8️⃣:숫자 기호
+ㅡ:9️⃣:숫자 기호
+ㅡ:0️⃣:숫자 기호
+ㅡ:🔟:숫자 기호
+ㅡ:🔢:숫자 기호
+ㅡ:#️⃣:기호
+ㅡ:🔣:기호
+ㅡ:⬆️:방향 기호
+ㅡ:⬇️:방향 기호
+ㅡ:⬅️:방향 기호
+ㅡ:➡️:방향 기호
+ㅡ:🔠:기호
+ㅡ:🔡:기호
+ㅡ:🔤:기호
+ㅡ:↗️:방향 기호
+ㅡ:↖️:방향 기호
+ㅡ:↘️:방향 기호
+ㅡ:↙️:방향 기호
+ㅡ:↔️:방향 기호
+ㅡ:↕️:방향 기호
+ㅡ:🔄:기호
+ㅡ:◀️:기호
+ㅡ:▶️:기호
+ㅡ:🔼:기호
+ㅡ:🔽:기호
+ㅡ:↩️:기호
+ㅡ:↪️:기호
+ㅡ:ℹ️:기호
+ㅡ:⏪:기호
+ㅡ:⏩:기호
+ㅡ:⏫:기호
+ㅡ:⏬:기호
+ㅡ:⤵️:기호
+ㅡ:⤴️:기호
+ㅡ:🆗:기호
+ㅡ:🔀:기호
+ㅡ:🔁:기호
+ㅡ:🔂:기호
+ㅡ:🆕:기호
+ㅡ:🆙:기호
+ㅡ:🆒:기호
+ㅡ:🆓:기호
+ㅡ:🆖:기호
+ㅡ:📶:기호
+ㅡ:🎦:기호
+ㅡ:🈁:기호
+ㅡ:🈯:기호
+ㅡ:🈳:기호
+ㅡ:🈵:기호
+ㅡ:🈴:기호
+ㅡ:🈲:기호
+ㅡ:🉐:기호
+ㅡ:🈹:기호
+ㅡ:🈺:기호
+ㅡ:🈶:기호
+ㅡ:🈚:기호
+ㅡ:🚻:기호
+ㅡ:🚹:기호
+ㅡ:🚺:기호
+ㅡ:🚼:기호
+ㅡ:🚾:기호
+ㅡ:🚰:기호
+ㅡ:🚮:기호
+ㅡ:🅿️:기호
+ㅡ:♿️:기호
+ㅡ:🚭:기호
+ㅡ:🈷:기호
+ㅡ:🈸:기호
+ㅡ:🈂:기호
+ㅡ:Ⓜ️:기호
+ㅡ:🛂:기호
+ㅡ:🛄:기호
+ㅡ:🛅:기호
+ㅡ:🛃:기호
+ㅡ:🉑:기호
+ㅡ:㊙️:기호
+ㅡ:㊗️:기호
+ㅡ:🆑:기호
+ㅡ:🆘:기호
+ㅡ:🆔:기호
+ㅡ:🚫:기호
+ㅡ:🔞:기호
+ㅡ:📵:기호
+ㅡ:🚯:기호
+ㅡ:🚱:기호
+ㅡ:🚳:기호
+ㅡ:🚷:기호
+ㅡ:🚸:기호
+ㅡ:⛔:기호
+ㅡ:✳️:기호
+ㅡ:❇️:기호
+ㅡ:❎:기호
+ㅡ:✅:기호
+ㅡ:✴️:기호
+ㅡ:💟:기호
+ㅡ:🆚:기호
+ㅡ:📳:기호
+ㅡ:📴:기호
+ㅡ:🅰:기호
+ㅡ:🅱:기호
+ㅡ:🆎:기호
+ㅡ:🅾:기호
+ㅡ:💠:기호
+ㅡ:➿:기호
+ㅡ:♻️:기호
+ㅡ:♈️:기호
+ㅡ:♉️:기호
+ㅡ:♊️:기호
+ㅡ:♋️:기호
+ㅡ:♌️:기호
+ㅡ:♍️:기호
+ㅡ:♎️:기호
+ㅡ:♏️:기호
+ㅡ:♐️:기호
+ㅡ:♑️:기호
+ㅡ:♒️:기호
+ㅡ:♓️:기호
+ㅡ:⛎:기호
+ㅡ:🔯:기호
+ㅡ:🏧:기호
+ㅡ:💹:기호
+ㅡ:💲:기호
+ㅡ:💱:기호
+ㅡ:©:기호
+ㅡ:®:기호
+ㅡ:™:기호
+ㅡ:❌:기호
+ㅡ:‼️:기호
+ㅡ:⁉️:기호
+ㅡ:❗:기호
+ㅡ:❓:기호
+ㅡ:❕:기호
+ㅡ:❔:기호
+ㅡ:⭕:기호
+ㅡ:🔝:기호
+ㅡ:🔚:기호
+ㅡ:🔙:기호
+ㅡ:🔛:기호
+ㅡ:🔜:기호
+ㅡ:🔃:기호
+ㅡ:🕛:열두시
+ㅡ:🕧:열두시 삼십분
+ㅡ:🕐:한시
+ㅡ:🕜:한시 삼십분
+ㅡ:🕑:두시
+ㅡ:🕝:두시 삼십분
+ㅡ:🕒:세시
+ㅡ:🕞:세시 삼십분
+ㅡ:🕓:네시
+ㅡ:🕟:네시 삼십분
+ㅡ:🕔:다섯시
+ㅡ:🕠:다섯시 삼십분
+ㅡ:🕕:여섯시
+ㅡ:🕖:일곱시
+ㅡ:🕗:여덟시
+ㅡ:🕘:아홉시
+ㅡ:🕙:열시
+ㅡ:🕚:열한시
+ㅡ:🕡:여섯시 삼십분
+ㅡ:🕢:일곱시 삼십분
+ㅡ:🕣:여덟시 삼십분
+ㅡ:🕤:아홉시 삼십분
+ㅡ:🕥:열시 삼십분
+ㅡ:🕦:열한시 삼십분
+ㅡ:✖️:곱셈
+ㅡ:➕:덧셈
+ㅡ:➖:뺄셈
+ㅡ:➗:나눗셈
+ㅡ:♠️:스페이드
+ㅡ:♥️:하트
+ㅡ:♣️:클럽
+ㅡ:♦️:다이아몬드
+ㅡ:💮:기호
+ㅡ:💯:백점
+ㅡ:✔️:기호
+ㅡ:☑️:기호
+ㅡ:🔘:기호
+ㅡ:🔗:기호
+ㅡ:➰:기호
+ㅡ:〰:기호
+ㅡ:〽️:기호
+ㅡ:🔱:기호
+ㅡ:◼️:기호
+ㅡ:◻️:기호
+ㅡ:️◾️:기호
+ㅡ:◽️:기호
+ㅡ:▪️:기호
+ㅡ:️▫️:기호
+ㅡ:🔺:기호
+ㅡ:🔲:기호
+ㅡ:🔳:기호
+ㅡ:⚫:기호
+ㅡ:⚪:기호
+ㅡ:🔴:기호
+ㅡ:🔵:기호
+ㅡ:🔻:기호
+ㅡ:⬜:기호
+ㅡ:⬛:기호
+ㅡ:🔶:기호
+ㅡ:🔷:기호
+ㅡ:🔸:기호
+ㅡ:🔹:기호
