+++ data/hanja/mssymbolr.txt
@@ -1,0 +1,987 @@
+ㄱ:　:
+ㄱ:！:
+ㄱ:＇:
+ㄱ:，:
+ㄱ:．:
+ㄱ:／:
+ㄱ:：:
+ㄱ:；:
+ㄱ:？:
+ㄱ:＾:
+ㄱ:＿:
+ㄱ:｀:
+ㄱ:｜:
+ㄱ:￣:
+ㄱ:、:
+ㄱ:。:
+ㄱ:·:
+ㄱ:‥:
+ㄱ:…:
+ㄱ:¨:
+ㄱ:〃:
+ㄱ:­:
+ㄱ:―:
+ㄱ:∥:
+ㄱ:＼:
+ㄱ:∼:
+ㄱ:´:
+ㄱ:～:
+ㄱ:ˇ:
+ㄱ:˘:
+ㄱ:˝:
+ㄱ:˚:
+ㄱ:˙:
+ㄱ:¸:
+ㄱ:˛:
+ㄱ:¡:
+ㄱ:¿:
+ㄱ:ː:
+ㄲ:Æ:
+ㄲ:Ð:
+ㄲ:Ħ:
+ㄲ:Ĳ:
+ㄲ:Ŀ:
+ㄲ:Ł:
+ㄲ:Ø:
+ㄲ:Œ:
+ㄲ:Þ:
+ㄲ:Ŧ:
+ㄲ:Ŋ:
+ㄲ:æ:
+ㄲ:đ:
+ㄲ:ð:
+ㄲ:ħ:
+ㄲ:ı:
+ㄲ:ĳ:
+ㄲ:ĸ:
+ㄲ:ŀ:
+ㄲ:ł:
+ㄲ:ø:
+ㄲ:œ:
+ㄲ:ß:
+ㄲ:þ:
+ㄲ:ŧ:
+ㄲ:ŋ:
+ㄲ:ŉ:
+ㄴ:＂:
+ㄴ:（:
+ㄴ:）:
+ㄴ:［:
+ㄴ:］:
+ㄴ:｛:
+ㄴ:｝:
+ㄴ:‘:
+ㄴ:’:
+ㄴ:“:
+ㄴ:”:
+ㄴ:〔:
+ㄴ:〕:
+ㄴ:〈:
+ㄴ:〉:
+ㄴ:《:
+ㄴ:》:
+ㄴ:「:
+ㄴ:」:
+ㄴ:『:
+ㄴ:』:
+ㄴ:【:
+ㄴ:】:
+ㄷ:＋:
+ㄷ:－:
+ㄷ:＜:
+ㄷ:＝:
+ㄷ:＞:
+ㄷ:±:
+ㄷ:×:
+ㄷ:÷:
+ㄷ:≠:
+ㄷ:≤:
+ㄷ:≥:
+ㄷ:∞:
+ㄷ:∴:
+ㄷ:♂:
+ㄷ:♀:
+ㄷ:∠:
+ㄷ:⊥:
+ㄷ:⌒:
+ㄷ:∂:
+ㄷ:∇:
+ㄷ:≡:
+ㄷ:≒:
+ㄷ:≪:
+ㄷ:≫:
+ㄷ:√:
+ㄷ:∽:
+ㄷ:∝:
+ㄷ:∵:
+ㄷ:∫:
+ㄷ:∬:
+ㄷ:∈:
+ㄷ:∋:
+ㄷ:⊆:
+ㄷ:⊇:
+ㄷ:⊂:
+ㄷ:⊃:
+ㄷ:∪:
+ㄷ:∩:
+ㄷ:∧:
+ㄷ:∨:
+ㄷ:￢:
+ㄷ:⇒:
+ㄷ:⇔:
+ㄷ:∀:
+ㄷ:∃:
+ㄷ:∮:
+ㄷ:∑:
+ㄷ:∏:
+ㄸ:ぁ:
+ㄸ:あ:
+ㄸ:ぃ:
+ㄸ:い:
+ㄸ:ぅ:
+ㄸ:う:
+ㄸ:ぇ:
+ㄸ:え:
+ㄸ:ぉ:
+ㄸ:お:
+ㄸ:か:
+ㄸ:が:
+ㄸ:き:
+ㄸ:ぎ:
+ㄸ:く:
+ㄸ:ぐ:
+ㄸ:け:
+ㄸ:げ:
+ㄸ:こ:
+ㄸ:ご:
+ㄸ:さ:
+ㄸ:ざ:
+ㄸ:し:
+ㄸ:じ:
+ㄸ:す:
+ㄸ:ず:
+ㄸ:せ:
+ㄸ:ぜ:
+ㄸ:そ:
+ㄸ:ぞ:
+ㄸ:た:
+ㄸ:だ:
+ㄸ:ち:
+ㄸ:ぢ:
+ㄸ:っ:
+ㄸ:つ:
+ㄸ:づ:
+ㄸ:て:
+ㄸ:で:
+ㄸ:と:
+ㄸ:ど:
+ㄸ:な:
+ㄸ:に:
+ㄸ:ぬ:
+ㄸ:ね:
+ㄸ:の:
+ㄸ:は:
+ㄸ:ば:
+ㄸ:ぱ:
+ㄸ:ひ:
+ㄸ:び:
+ㄸ:ぴ:
+ㄸ:ふ:
+ㄸ:ぶ:
+ㄸ:ぷ:
+ㄸ:へ:
+ㄸ:べ:
+ㄸ:ぺ:
+ㄸ:ほ:
+ㄸ:ぼ:
+ㄸ:ぽ:
+ㄸ:ま:
+ㄸ:み:
+ㄸ:む:
+ㄸ:め:
+ㄸ:も:
+ㄸ:ゃ:
+ㄸ:や:
+ㄸ:ゅ:
+ㄸ:ゆ:
+ㄸ:ょ:
+ㄸ:よ:
+ㄸ:ら:
+ㄸ:り:
+ㄸ:る:
+ㄸ:れ:
+ㄸ:ろ:
+ㄸ:ゎ:
+ㄸ:わ:
+ㄸ:ゐ:
+ㄸ:ゑ:
+ㄸ:を:
+ㄸ:ん:
+ㄹ:＄:
+ㄹ:％:
+ㄹ:￦:
+ㄹ:Ｆ:
+ㄹ:′:
+ㄹ:″:
+ㄹ:℃:
+ㄹ:Å:
+ㄹ:￠:
+ㄹ:￡:
+ㄹ:￥:
+ㄹ:¤:
+ㄹ:℉:
+ㄹ:‰:
+ㄹ:€:
+ㄹ:㎕:
+ㄹ:㎖:
+ㄹ:㎗:
+ㄹ:ℓ:
+ㄹ:㎘:
+ㄹ:㏄:
+ㄹ:㎣:
+ㄹ:㎤:
+ㄹ:㎥:
+ㄹ:㎦:
+ㄹ:㎙:
+ㄹ:㎚:
+ㄹ:㎛:
+ㄹ:㎜:
+ㄹ:㎝:
+ㄹ:㎞:
+ㄹ:㎟:
+ㄹ:㎠:
+ㄹ:㎡:
+ㄹ:㎢:
+ㄹ:㏊:
+ㄹ:㎍:
+ㄹ:㎎:
+ㄹ:㎏:
+ㄹ:㏏:
+ㄹ:㎈:
+ㄹ:㎉:
+ㄹ:㏈:
+ㄹ:㎧:
+ㄹ:㎨:
+ㄹ:㎰:
+ㄹ:㎱:
+ㄹ:㎲:
+ㄹ:㎳:
+ㄹ:㎴:
+ㄹ:㎵:
+ㄹ:㎶:
+ㄹ:㎷:
+ㄹ:㎸:
+ㄹ:㎹:
+ㄹ:㎀:
+ㄹ:㎁:
+ㄹ:㎂:
+ㄹ:㎃:
+ㄹ:㎄:
+ㄹ:㎺:
+ㄹ:㎻:
+ㄹ:㎼:
+ㄹ:㎽:
+ㄹ:㎾:
+ㄹ:㎿:
+ㄹ:㎐:
+ㄹ:㎑:
+ㄹ:㎒:
+ㄹ:㎓:
+ㄹ:㎔:
+ㄹ:Ω:
+ㄹ:㏀:
+ㄹ:㏁:
+ㄹ:㎊:
+ㄹ:㎋:
+ㄹ:㎌:
+ㄹ:㏖:
+ㄹ:㏅:
+ㄹ:㎭:
+ㄹ:㎮:
+ㄹ:㎯:
+ㄹ:㏛:
+ㄹ:㎩:
+ㄹ:㎪:
+ㄹ:㎫:
+ㄹ:㎬:
+ㄹ:㏝:
+ㄹ:㏐:
+ㄹ:㏓:
+ㄹ:㏃:
+ㄹ:㏉:
+ㄹ:㏜:
+ㄹ:㏆:
+ㅁ:＃:
+ㅁ:＆:
+ㅁ:＊:
+ㅁ:＠:
+ㅁ:§:
+ㅁ:※:
+ㅁ:☆:
+ㅁ:★:
+ㅁ:○:
+ㅁ:●:
+ㅁ:◎:
+ㅁ:◇:
+ㅁ:◆:
+ㅁ:□:
+ㅁ:■:
+ㅁ:△:
+ㅁ:▲:
+ㅁ:▽:
+ㅁ:▼:
+ㅁ:→:
+ㅁ:←:
+ㅁ:↑:
+ㅁ:↓:
+ㅁ:↔:
+ㅁ:〓:
+ㅁ:◁:
+ㅁ:◀:
+ㅁ:▷:
+ㅁ:▶:
+ㅁ:♤:
+ㅁ:♠:
+ㅁ:♡:
+ㅁ:♥:
+ㅁ:♧:
+ㅁ:♣:
+ㅁ:⊙:
+ㅁ:◈:
+ㅁ:▣:
+ㅁ:◐:
+ㅁ:◑:
+ㅁ:▒:
+ㅁ:▤:
+ㅁ:▥:
+ㅁ:▨:
+ㅁ:▧:
+ㅁ:▦:
+ㅁ:▩:
+ㅁ:♨:
+ㅁ:☏:
+ㅁ:☎:
+ㅁ:☜:
+ㅁ:☞:
+ㅁ:¶:
+ㅁ:†:
+ㅁ:‡:
+ㅁ:↕:
+ㅁ:↗:
+ㅁ:↙:
+ㅁ:↖:
+ㅁ:↘:
+ㅁ:♭:
+ㅁ:♩:
+ㅁ:♪:
+ㅁ:♬:
+ㅁ:㉿:
+ㅁ:㈜:
+ㅁ:№:
+ㅁ:㏇:
+ㅁ:™:
+ㅁ:㏂:
+ㅁ:㏘:
+ㅁ:℡:
+ㅁ:®:
+ㅁ:ª:
+ㅁ:º:
+ㅂ:─:
+ㅂ:│:
+ㅂ:┌:
+ㅂ:┐:
+ㅂ:┘:
+ㅂ:└:
+ㅂ:├:
+ㅂ:┬:
+ㅂ:┤:
+ㅂ:┴:
+ㅂ:┼:
+ㅂ:━:
+ㅂ:┃:
+ㅂ:┏:
+ㅂ:┓:
+ㅂ:┛:
+ㅂ:┗:
+ㅂ:┣:
+ㅂ:┳:
+ㅂ:┫:
+ㅂ:┻:
+ㅂ:╋:
+ㅂ:┠:
+ㅂ:┯:
+ㅂ:┨:
+ㅂ:┷:
+ㅂ:┿:
+ㅂ:┝:
+ㅂ:┰:
+ㅂ:┥:
+ㅂ:┸:
+ㅂ:╂:
+ㅂ:┒:
+ㅂ:┑:
+ㅂ:┚:
+ㅂ:┙:
+ㅂ:┖:
+ㅂ:┕:
+ㅂ:┎:
+ㅂ:┍:
+ㅂ:┞:
+ㅂ:┟:
+ㅂ:┡:
+ㅂ:┢:
+ㅂ:┦:
+ㅂ:┧:
+ㅂ:┩:
+ㅂ:┪:
+ㅂ:┭:
+ㅂ:┮:
+ㅂ:┱:
+ㅂ:┲:
+ㅂ:┵:
+ㅂ:┶:
+ㅂ:┹:
+ㅂ:┺:
+ㅂ:┽:
+ㅂ:┾:
+ㅂ:╀:
+ㅂ:╁:
+ㅂ:╃:
+ㅂ:╄:
+ㅂ:╅:
+ㅂ:╆:
+ㅂ:╇:
+ㅂ:╈:
+ㅂ:╉:
+ㅂ:╊:
+ㅃ:ァ:
+ㅃ:ア:
+ㅃ:ィ:
+ㅃ:イ:
+ㅃ:ゥ:
+ㅃ:ウ:
+ㅃ:ェ:
+ㅃ:エ:
+ㅃ:ォ:
+ㅃ:オ:
+ㅃ:カ:
+ㅃ:ガ:
+ㅃ:キ:
+ㅃ:ギ:
+ㅃ:ク:
+ㅃ:グ:
+ㅃ:ケ:
+ㅃ:ゲ:
+ㅃ:コ:
+ㅃ:ゴ:
+ㅃ:サ:
+ㅃ:ザ:
+ㅃ:シ:
+ㅃ:ジ:
+ㅃ:ス:
+ㅃ:ズ:
+ㅃ:セ:
+ㅃ:ゼ:
+ㅃ:ソ:
+ㅃ:ゾ:
+ㅃ:タ:
+ㅃ:ダ:
+ㅃ:チ:
+ㅃ:ヂ:
+ㅃ:ッ:
+ㅃ:ツ:
+ㅃ:ヅ:
+ㅃ:テ:
+ㅃ:デ:
+ㅃ:ト:
+ㅃ:ド:
+ㅃ:ナ:
+ㅃ:ニ:
+ㅃ:ヌ:
+ㅃ:ネ:
+ㅃ:ノ:
+ㅃ:ハ:
+ㅃ:バ:
+ㅃ:パ:
+ㅃ:ヒ:
+ㅃ:ビ:
+ㅃ:ピ:
+ㅃ:フ:
+ㅃ:ブ:
+ㅃ:プ:
+ㅃ:ヘ:
+ㅃ:ベ:
+ㅃ:ペ:
+ㅃ:ホ:
+ㅃ:ボ:
+ㅃ:ポ:
+ㅃ:マ:
+ㅃ:ミ:
+ㅃ:ム:
+ㅃ:メ:
+ㅃ:モ:
+ㅃ:ャ:
+ㅃ:ヤ:
+ㅃ:ュ:
+ㅃ:ユ:
+ㅃ:ョ:
+ㅃ:ヨ:
+ㅃ:ラ:
+ㅃ:リ:
+ㅃ:ル:
+ㅃ:レ:
+ㅃ:ロ:
+ㅃ:ヮ:
+ㅃ:ワ:
+ㅃ:ヰ:
+ㅃ:ヱ:
+ㅃ:ヲ:
+ㅃ:ン:
+ㅃ:ヴ:
+ㅃ:ヵ:
+ㅃ:ヶ:
+ㅅ:㉠:
+ㅅ:㉡:
+ㅅ:㉢:
+ㅅ:㉣:
+ㅅ:㉤:
+ㅅ:㉥:
+ㅅ:㉦:
+ㅅ:㉧:
+ㅅ:㉨:
+ㅅ:㉩:
+ㅅ:㉪:
+ㅅ:㉫:
+ㅅ:㉬:
+ㅅ:㉭:
+ㅅ:㉮:
+ㅅ:㉯:
+ㅅ:㉰:
+ㅅ:㉱:
+ㅅ:㉲:
+ㅅ:㉳:
+ㅅ:㉴:
+ㅅ:㉵:
+ㅅ:㉶:
+ㅅ:㉷:
+ㅅ:㉸:
+ㅅ:㉹:
+ㅅ:㉺:
+ㅅ:㉻:
+ㅅ:㈀:
+ㅅ:㈁:
+ㅅ:㈂:
+ㅅ:㈃:
+ㅅ:㈄:
+ㅅ:㈅:
+ㅅ:㈆:
+ㅅ:㈇:
+ㅅ:㈈:
+ㅅ:㈉:
+ㅅ:㈊:
+ㅅ:㈋:
+ㅅ:㈌:
+ㅅ:㈍:
+ㅅ:㈎:
+ㅅ:㈏:
+ㅅ:㈐:
+ㅅ:㈑:
+ㅅ:㈒:
+ㅅ:㈓:
+ㅅ:㈔:
+ㅅ:㈕:
+ㅅ:㈖:
+ㅅ:㈗:
+ㅅ:㈘:
+ㅅ:㈙:
+ㅅ:㈚:
+ㅅ:㈛:
+ㅆ:А:
+ㅆ:Б:
+ㅆ:В:
+ㅆ:Г:
+ㅆ:Д:
+ㅆ:Е:
+ㅆ:Ё:
+ㅆ:Ж:
+ㅆ:З:
+ㅆ:И:
+ㅆ:Й:
+ㅆ:К:
+ㅆ:Л:
+ㅆ:М:
+ㅆ:Н:
+ㅆ:О:
+ㅆ:П:
+ㅆ:Р:
+ㅆ:С:
+ㅆ:Т:
+ㅆ:У:
+ㅆ:Ф:
+ㅆ:Х:
+ㅆ:Ц:
+ㅆ:Ч:
+ㅆ:Ш:
+ㅆ:Щ:
+ㅆ:Ъ:
+ㅆ:Ы:
+ㅆ:Ь:
+ㅆ:Э:
+ㅆ:Ю:
+ㅆ:Я:
+ㅆ:а:
+ㅆ:б:
+ㅆ:в:
+ㅆ:г:
+ㅆ:д:
+ㅆ:е:
+ㅆ:ё:
+ㅆ:ж:
+ㅆ:з:
+ㅆ:и:
+ㅆ:й:
+ㅆ:к:
+ㅆ:л:
+ㅆ:м:
+ㅆ:н:
+ㅆ:о:
+ㅆ:п:
+ㅆ:р:
+ㅆ:с:
+ㅆ:т:
+ㅆ:у:
+ㅆ:ф:
+ㅆ:х:
+ㅆ:ц:
+ㅆ:ч:
+ㅆ:ш:
+ㅆ:щ:
+ㅆ:ъ:
+ㅆ:ы:
+ㅆ:ь:
+ㅆ:э:
+ㅆ:ю:
+ㅆ:я:
+ㅇ:ⓐ:
+ㅇ:ⓑ:
+ㅇ:ⓒ:
+ㅇ:ⓓ:
+ㅇ:ⓔ:
+ㅇ:ⓕ:
+ㅇ:ⓖ:
+ㅇ:ⓗ:
+ㅇ:ⓘ:
+ㅇ:ⓙ:
+ㅇ:ⓚ:
+ㅇ:ⓛ:
+ㅇ:ⓜ:
+ㅇ:ⓝ:
+ㅇ:ⓞ:
+ㅇ:ⓟ:
+ㅇ:ⓠ:
+ㅇ:ⓡ:
+ㅇ:ⓢ:
+ㅇ:ⓣ:
+ㅇ:ⓤ:
+ㅇ:ⓥ:
+ㅇ:ⓦ:
+ㅇ:ⓧ:
+ㅇ:ⓨ:
+ㅇ:ⓩ:
+ㅇ:①:
+ㅇ:②:
+ㅇ:③:
+ㅇ:④:
+ㅇ:⑤:
+ㅇ:⑥:
+ㅇ:⑦:
+ㅇ:⑧:
+ㅇ:⑨:
+ㅇ:⑩:
+ㅇ:⑪:
+ㅇ:⑫:
+ㅇ:⑬:
+ㅇ:⑭:
+ㅇ:⑮:
+ㅇ:⒜:
+ㅇ:⒝:
+ㅇ:⒞:
+ㅇ:⒟:
+ㅇ:⒠:
+ㅇ:⒡:
+ㅇ:⒢:
+ㅇ:⒣:
+ㅇ:⒤:
+ㅇ:⒥:
+ㅇ:⒦:
+ㅇ:⒧:
+ㅇ:⒨:
+ㅇ:⒩:
+ㅇ:⒪:
+ㅇ:⒫:
+ㅇ:⒬:
+ㅇ:⒭:
+ㅇ:⒮:
+ㅇ:⒯:
+ㅇ:⒰:
+ㅇ:⒱:
+ㅇ:⒲:
+ㅇ:⒳:
+ㅇ:⒴:
+ㅇ:⒵:
+ㅇ:⑴:
+ㅇ:⑵:
+ㅇ:⑶:
+ㅇ:⑷:
+ㅇ:⑸:
+ㅇ:⑹:
+ㅇ:⑺:
+ㅇ:⑻:
+ㅇ:⑼:
+ㅇ:⑽:
+ㅇ:⑾:
+ㅇ:⑿:
+ㅇ:⒀:
+ㅇ:⒁:
+ㅇ:⒂:
+ㅈ:０:
+ㅈ:１:
+ㅈ:２:
+ㅈ:３:
+ㅈ:４:
+ㅈ:５:
+ㅈ:６:
+ㅈ:７:
+ㅈ:８:
+ㅈ:９:
+ㅈ:ⅰ:
+ㅈ:ⅱ:
+ㅈ:ⅲ:
+ㅈ:ⅳ:
+ㅈ:ⅴ:
+ㅈ:ⅵ:
+ㅈ:ⅶ:
+ㅈ:ⅷ:
+ㅈ:ⅸ:
+ㅈ:ⅹ:
+ㅈ:Ⅰ:
+ㅈ:Ⅱ:
+ㅈ:Ⅲ:
+ㅈ:Ⅳ:
+ㅈ:Ⅴ:
+ㅈ:Ⅵ:
+ㅈ:Ⅶ:
+ㅈ:Ⅷ:
+ㅈ:Ⅸ:
+ㅈ:Ⅹ:
+ㅊ:½:
+ㅊ:⅓:
+ㅊ:⅔:
+ㅊ:¼:
+ㅊ:¾:
+ㅊ:⅛:
+ㅊ:⅜:
+ㅊ:⅝:
+ㅊ:⅞:
+ㅊ:¹:
+ㅊ:²:
+ㅊ:³:
+ㅊ:⁴:
+ㅊ:ⁿ:
+ㅊ:₁:
+ㅊ:₂:
+ㅊ:₃:
+ㅊ:₄:
+ㅋ:ㄱ:
+ㅋ:ㄲ:
+ㅋ:ㄳ:
+ㅋ:ㄴ:
+ㅋ:ㄵ:
+ㅋ:ㄶ:
+ㅋ:ㄷ:
+ㅋ:ㄸ:
+ㅋ:ㄹ:
+ㅋ:ㄺ:
+ㅋ:ㄻ:
+ㅋ:ㄼ:
+ㅋ:ㄽ:
+ㅋ:ㄾ:
+ㅋ:ㄿ:
+ㅋ:ㅀ:
+ㅋ:ㅁ:
+ㅋ:ㅂ:
+ㅋ:ㅃ:
+ㅋ:ㅄ:
+ㅋ:ㅅ:
+ㅋ:ㅆ:
+ㅋ:ㅇ:
+ㅋ:ㅈ:
+ㅋ:ㅉ:
+ㅋ:ㅊ:
+ㅋ:ㅋ:
+ㅋ:ㅌ:
+ㅋ:ㅍ:
+ㅋ:ㅎ:
+ㅋ:ㅏ:
+ㅋ:ㅐ:
+ㅋ:ㅑ:
+ㅋ:ㅒ:
+ㅋ:ㅓ:
+ㅋ:ㅔ:
+ㅋ:ㅕ:
+ㅋ:ㅖ:
+ㅋ:ㅗ:
+ㅋ:ㅘ:
+ㅋ:ㅙ:
+ㅋ:ㅚ:
+ㅋ:ㅛ:
+ㅋ:ㅜ:
+ㅋ:ㅝ:
+ㅋ:ㅞ:
+ㅋ:ㅟ:
+ㅋ:ㅠ:
+ㅋ:ㅡ:
+ㅋ:ㅢ:
+ㅋ:ㅣ:
+ㅌ:ㅥ:
+ㅌ:ㅦ:
+ㅌ:ㅧ:
+ㅌ:ㅨ:
+ㅌ:ㅩ:
+ㅌ:ㅪ:
+ㅌ:ㅫ:
+ㅌ:ㅬ:
+ㅌ:ㅭ:
+ㅌ:ㅮ:
+ㅌ:ㅯ:
+ㅌ:ㅰ:
+ㅌ:ㅱ:
+ㅌ:ㅲ:
+ㅌ:ㅳ:
+ㅌ:ㅴ:
+ㅌ:ㅵ:
+ㅌ:ㅶ:
+ㅌ:ㅷ:
+ㅌ:ㅸ:
+ㅌ:ㅹ:
+ㅌ:ㅺ:
+ㅌ:ㅻ:
+ㅌ:ㅼ:
+ㅌ:ㅽ:
+ㅌ:ㅾ:
+ㅌ:ㅿ:
+ㅌ:ㆀ:
+ㅌ:ㆁ:
+ㅌ:ㆂ:
+ㅌ:ㆃ:
+ㅌ:ㆄ:
+ㅌ:ㆅ:
+ㅌ:ㆆ:
+ㅌ:ㆇ:
+ㅌ:ㆈ:
+ㅌ:ㆉ:
+ㅌ:ㆊ:
+ㅌ:ㆋ:
+ㅌ:ㆌ:
+ㅌ:ㆍ:
+ㅌ:ㆎ:
+ㅍ:Ａ:
+ㅍ:Ｂ:
+ㅍ:Ｃ:
+ㅍ:Ｄ:
+ㅍ:Ｅ:
+ㅍ:Ｆ:
+ㅍ:Ｇ:
+ㅍ:Ｈ:
+ㅍ:Ｉ:
+ㅍ:Ｊ:
+ㅍ:Ｋ:
+ㅍ:Ｌ:
+ㅍ:Ｍ:
+ㅍ:Ｎ:
+ㅍ:Ｏ:
+ㅍ:Ｐ:
+ㅍ:Ｑ:
+ㅍ:Ｒ:
+ㅍ:Ｓ:
+ㅍ:Ｔ:
+ㅍ:Ｕ:
+ㅍ:Ｖ:
+ㅍ:Ｗ:
+ㅍ:Ｘ:
+ㅍ:Ｙ:
+ㅍ:Ｚ:
+ㅍ:ａ:
+ㅍ:ｂ:
+ㅍ:ｃ:
+ㅍ:ｄ:
+ㅍ:ｅ:
+ㅍ:ｆ:
+ㅍ:ｇ:
+ㅍ:ｈ:
+ㅍ:ｉ:
+ㅍ:ｊ:
+ㅍ:ｋ:
+ㅍ:ｌ:
+ㅍ:ｍ:
+ㅍ:ｎ:
+ㅍ:ｏ:
+ㅍ:ｐ:
+ㅍ:ｑ:
+ㅍ:ｒ:
+ㅍ:ｓ:
+ㅍ:ｔ:
+ㅍ:ｕ:
+ㅍ:ｖ:
+ㅍ:ｗ:
+ㅍ:ｘ:
+ㅍ:ｙ:
+ㅍ:ｚ:
+ㅎ:Α:
+ㅎ:Β:
+ㅎ:Γ:
+ㅎ:Δ:
+ㅎ:Ε:
+ㅎ:Ζ:
+ㅎ:Η:
+ㅎ:Θ:
+ㅎ:Ι:
+ㅎ:Κ:
+ㅎ:Λ:
+ㅎ:Μ:
+ㅎ:Ν:
+ㅎ:Ξ:
+ㅎ:Ο:
+ㅎ:Π:
+ㅎ:Ρ:
+ㅎ:Σ:
+ㅎ:Τ:
+ㅎ:Υ:
+ㅎ:Φ:
+ㅎ:Χ:
+ㅎ:Ψ:
+ㅎ:Ω:
+ㅎ:α:
+ㅎ:β:
+ㅎ:γ:
+ㅎ:δ:
+ㅎ:ε:
+ㅎ:ζ:
+ㅎ:η:
+ㅎ:θ:
+ㅎ:ι:
+ㅎ:κ:
+ㅎ:λ:
+ㅎ:μ:
+ㅎ:ν:
+ㅎ:ξ:
+ㅎ:ο:
+ㅎ:π:
+ㅎ:ρ:
+ㅎ:σ:
+ㅎ:τ:
+ㅎ:υ:
+ㅎ:φ:
+ㅎ:χ:
+ㅎ:ψ:
+ㅎ:ω:
